move processMultiUpdateAfter function to FlinkChangeTaskWriter
