+++ core/src/main/java/com/netease/arctic/io/writer/BaseTaskWriter.java
@@ -123,6 +123,10 @@ public abstract class BaseTaskWriter<T> implements TaskWriter<T> {
     writerHolder.close();
   }
 
+  public PrimaryKeyData getPrimaryKey() {
+    return primaryKey;
+  }
+
   /**
    * Wrap the data as a {@link StructLike}.
    */
