+++ flink/v1.12/flink/src/main/java/com/netease/arctic/flink/write/FlinkChangeTaskWriter.java
@@ -19,6 +19,7 @@
 package com.netease.arctic.flink.write;
 
 import com.netease.arctic.data.ChangeAction;
+import com.netease.arctic.data.PrimaryKeyData;
 import com.netease.arctic.io.ArcticFileIO;
 import com.netease.arctic.io.writer.ChangeTaskWriter;
 import com.netease.arctic.io.writer.OutputFileFactory;
@@ -28,12 +29,17 @@ import org.apache.flink.table.data.GenericRowData;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.data.utils.JoinedRowData;
 import org.apache.flink.table.types.logical.RowType;
+import org.apache.flink.types.RowKind;
 import org.apache.iceberg.FileFormat;
 import org.apache.iceberg.PartitionSpec;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.StructLike;
 import org.apache.iceberg.flink.RowDataWrapper;
 import org.apache.iceberg.io.FileAppenderFactory;
+
+import java.io.IOException;
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * task writer for {@link KeyedTable#changeTable()} ()}.
@@ -42,14 +48,17 @@ import org.apache.iceberg.io.FileAppenderFactory;
 public class FlinkChangeTaskWriter extends ChangeTaskWriter<RowData> {
 
   private final RowDataWrapper wrapper;
+  private final boolean upsert;
+  private Set<PrimaryKeyData> hasUpdateBeforeKeys = new HashSet<>();
 
   public FlinkChangeTaskWriter(FileFormat format, FileAppenderFactory<RowData> appenderFactory,
                                OutputFileFactory outputFileFactory, ArcticFileIO io, long targetFileSize,
                                long mask, Schema schema, RowType flinkSchema, PartitionSpec spec,
-                               PrimaryKeySpec primaryKeySpec) {
+                               PrimaryKeySpec primaryKeySpec, boolean upsert) {
     super(format, appenderFactory, outputFileFactory, io,
         targetFileSize, mask, schema, spec, primaryKeySpec, false);
     this.wrapper = new RowDataWrapper(flinkSchema, schema.asStruct());
+    this.upsert = upsert;
   }
 
   @Override
@@ -60,6 +69,17 @@ public class FlinkChangeTaskWriter extends ChangeTaskWriter<RowData> {
   @Override
   protected RowData appendMetaColumns(RowData data, Long fileOffset) {
     return new JoinedRowData(data, GenericRowData.of(fileOffset));
+  }
+
+  @Override
+  public void write(RowData row) throws IOException {
+    processMultiUpdateAfter(row);
+    if (upsert && RowKind.INSERT.equals(row.getRowKind())) {
+      row.setRowKind(RowKind.DELETE);
+      super.write(row);
+      row.setRowKind(RowKind.INSERT);
+    }
+    super.write(row);
   }
 
   @Override
@@ -75,5 +95,23 @@ public class FlinkChangeTaskWriter extends ChangeTaskWriter<RowData> {
         return ChangeAction.UPDATE_AFTER;
     }
     return ChangeAction.INSERT;
+  }
+
+  /**
+   * Turn update_after to insert if there isn't update_after followed by update_before.
+   */
+  private void processMultiUpdateAfter(RowData row) {
+    PrimaryKeyData primaryKey = super.getPrimaryKey();
+    primaryKey.primaryKey(wrapper.wrap(row));
+    PrimaryKeyData copyKey = primaryKey.copy();
+
+    if (RowKind.UPDATE_AFTER.equals(row.getRowKind()) && !hasUpdateBeforeKeys.contains(copyKey)) {
+      row.setRowKind(RowKind.INSERT);
+    }
+    if (RowKind.UPDATE_BEFORE.equals(row.getRowKind())) {
+      hasUpdateBeforeKeys.add(copyKey);
+    } else {
+      hasUpdateBeforeKeys.remove(copyKey);
+    }
   }
 }
